老生常谈的问题，虚焊。早上用电表测的时候没发现虚焊，但是是在没办法了用烙铁拖了一下，结果还真就好了，今天测OV5640得知一个方法，就是让摄像头区照红绿蓝三张纯色图片，这样就能判断哪里出问题了——大W

+++ firmware/STM32H743core/Core/Inc/lcd.h
@@ -21,7 +21,7 @@ uint32_t mypow(uint8_t m,uint8_t n);//求幂
 void LCD_ShowIntNum(uint16_t x,uint16_t y,uint16_t num,uint8_t len,uint16_t fc,uint16_t bc,uint8_t sizey);//显示整数变量
 void LCD_ShowFloatNum1(uint16_t x,uint16_t y,float num,uint8_t len,uint16_t fc,uint16_t bc,uint8_t sizey);//显示两位小数变量
 
-void LCD_ShowPicture(uint16_t x,uint16_t y,uint16_t length,uint16_t width,const uint8_t pic[]);//显示图片 //172,320,172,320
+void LCD_ShowPicture(uint16_t x,uint16_t y,uint16_t length,uint16_t width,const uint16_t pic[]);//显示图片 //172,320,172,320
 
 
 //画笔颜色
